backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
pymysql==1.1.0
cryptography==41.0.7
python-decouple==3.8
Pillow==10.2.0
gunicorn==20.1.0
whitenoise==6.5.0
